# update gogo protobuf, generate line