--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,7 @@ module github.com/paulmach/osm
 go 1.13
 
 require (
-	github.com/gogo/protobuf v1.3.0
-	github.com/paulmach/orb v0.1.5
+	github.com/gogo/protobuf v1.3.1
+	github.com/paulmach/orb v0.1.6
 	golang.org/x/time v0.0.0-20190921001708-c4c64cad1fd0
 )