--- a/osmpbf/internal/osmpbf/generate.go
+++ b/osmpbf/internal/osmpbf/generate.go
@@ -2,4 +2,4 @@ package osmpbf
 
 // gogo protobuf seems to be about 10% faster
 // https://github.com/gogo/protobuf
-//go:generate protoc --proto_path=../../../../../..:../../../../../gogo/protobuf/protobuf:. --gogofaster_out=. fileformat.proto osmformat.proto
+//go:generate protoc --proto_path=$GOPATH/src/:$GOPATH/src/github.com/gogo/protobuf/protobuf:. --gogofaster_out=.  fileformat.proto osmformat.proto