Updating gogo protobuf gave me an 15% performance increase for parsing large pbf files.