@paulmach 2020-06-22
nice!
```
benchmark             old ns/op     new ns/op     delta
BenchmarkDecode-4     962648671     868660926     -9.76%

benchmark             old allocs     new allocs     delta
BenchmarkDecode-4     9730439        7732791        -20.53%

benchmark             old bytes      new bytes      delta
BenchmarkDecode-4     1916298720     1306693312     -31.81%
```